hal/pmc: refactor into clocks mod, split clock impls
This commit refactors the pmc module into a more general clocks
module (because what pmc did previously covered both §30 and §31),
splits the implementations into their own files for improved
readability and exhausively documents the new API.

In comparison with the old implementation, this one is actually
token (singleton)-based: the compiler will now complain if the user
tries to configure a clock that has already been configured, because
the token has been moved/consumed.

Further, the user is now forced to configure the whole clock
hierarchy: the HostClock that is required to power peripherals is
acquired only when the HostClockController has been configured.

Fixed in this commit is also a UpllClock dependency: before it was
possible to configure this clock with an InternalRC-configured
SlowClock. This is invalid, it is now only possible to configure the
clock with a SlowClock<ExternalNormal> or SlowClock<ExternalBypass>,
as per Figure 31-1.

Additionally, Pmc::enable_peripherals (now
HostClock::enable_peripheral) has been greatly simplified.