hal/serial/spi: fix chip select
Before this commit, it was presumed that the memory layout offset
index of NPCSx also acted as the SPI_TDR.PCS value for the same NPCSx.
This was only the case for NPCS0 and NPCS1. For NPCS2, NPCS0 would be
selected instead; and for NPCS3, NPCS2.

Now, the correct hardware descriptor is written to PCS, and all four
NPCSx are properly selected.