hal/pio: Pin::into_mode: switch PB4/5/6/7/12 from alternate mode
Before this commit, it was not possible to use these pins as I/Os.
This is now possible if the reconfigurable-system-pins feature is
enabled.

Due to the general purpose the bus matrix serves (more than just PIO),
and because it is not used elsewhere, we unsafely access it instead of
consuming the pac::MATRIX. If we were to consume it in the API, the
API would have to be broken again when we find a better place to
consume it.

Closes #41.