pio: implement open-drain/multi-drive API
This patch implements the open drain ("multi-driver") mode with
pull-up and pull-down for each pin.

I could not find any information indicating that the modes are not
supported for any specific pins and have as such followed the same
patterns of implementation as for the other pin modes.

Additionally, the InputPin and OutputPin traits are implemented for
OpenDrain pins. This permits using crates like onewire to quickly
implement software 1-wire support for Dallas-style sensors (ds18b20
for instance).